Refactor : 일관성 유지를 위해 package 만들기
